requirements.txt
fastapi
uvicorn
requests
streamlit
python-dotenv
openai
